requirements.txt
aiofiles~=0.5.0
fastapi[all]~=0.62.0
jinja2~=2.11.2
pydantic~=1.7.3
pyyaml~=5.3.1
starlette~=0.13.6